2002-12-27 - PONG.ASM was written by me and a friend of mine on a Friday afternoon in 2002
The earliest timestamp I have on this file is 2002-12-27 23:01

It was coded on a 486 with intel x86 instructions.

We had originally planned on making it into an interactive game, but never got around to it.

As it is now (and as we left it), a ball just bounces around the screen.